requirements.txt
llama-index==0.10.33
openai==1.98.0
python-dotenv==1.0.1
streamlit==1.35.0
markdownify==0.11.6
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-dotenv==1.0.1
